2nd_sprint was merged, code clean up and refactoring was made
