+++ app/src/main/java/com/example/quizer/MainActivity.java
@@ -4,10 +4,8 @@ import android.content.SharedPreferences;
 
 import androidx.fragment.app.Fragment;
 
-import com.example.quizer.recyclerView.ListFragmentActivity;
 import com.example.quizer.recyclerView.QuizListFragment;
 import com.example.quizer.userCabinet.AuthorizationFragment;
-import com.example.quizer.userCabinet.RegistrationFragment;
 
 
 public class MainActivity extends SingleFragmentActivity {
@@ -18,7 +16,7 @@ public class MainActivity extends SingleFragmentActivity {
         SharedPreferences preferences = getSharedPreferences(USER_ID_PREF, 0);
         int userId = preferences.getInt(USER_ID_PREF,-1);
         if(userId==-1)
-            return new RegistrationFragment();
+            return new AuthorizationFragment();
             //return new AuthorizationFragment();
         else return new QuizListFragment();
