+++ app/src/main/java/com/example/quizer/rest/UserAPI.java
@@ -1,11 +1,10 @@
-package com.example.quizer.database;
+package com.example.quizer.rest;
 
 import com.example.quizer.userCabinet.User;
 
 
 import retrofit2.Call;
 import retrofit2.http.Body;
-import retrofit2.http.Headers;
 import retrofit2.http.POST;
 
 public interface UserAPI {
@@ -13,7 +12,4 @@ public interface UserAPI {
     @POST("/rest/user/authorize")
     Call<Integer> authorizeUser(@Body User user );
 
-    @Headers("Content-Type: application/json")
-    @POST("/rest/user/register")
-    Call<Integer> registerUser(@Body User user);
 }
