--- app/src/main/java/com/example/quizer/ui/main/SectionsPagerAdapter.java
@@ -1,51 +1,0 @@
-package com.example.quizer.ui.main;
-
-import android.content.Context;
-
-import androidx.annotation.Nullable;
-import androidx.annotation.StringRes;
-import androidx.fragment.app.Fragment;
-import androidx.fragment.app.FragmentManager;
-import androidx.fragment.app.FragmentPagerAdapter;
-
-import com.example.quizer.R;
-import com.example.quizer.userCabinet.AuthorizationFragment;
-import com.example.quizer.userCabinet.RegistrationFragment;
-
-/**
- * A [FragmentPagerAdapter] that returns a fragment corresponding to
- * one of the sections/tabs/pages.
- */
-public class SectionsPagerAdapter extends FragmentPagerAdapter {
-
-    @StringRes
-    private static final int[] TAB_TITLES = new int[]{R.string.tab_text_1, R.string.tab_text_2};
-    private final Context mContext;
-
-    public SectionsPagerAdapter(Context context, FragmentManager fm) {
-        super(fm);
-        mContext = context;
-    }
-
-    @Override
-    public Fragment getItem(int position) {
-        switch (position) {
-            case 0:
-                return new RegistrationFragment();
-            case 1:
-                return new AuthorizationFragment();
-        }
-        return PlaceholderFragment.newInstance(position + 1);
-    }
-
-    @Nullable
-    @Override
-    public CharSequence getPageTitle(int position) {
-        return mContext.getResources().getString(TAB_TITLES[position]);
-    }
-
-    @Override
-    public int getCount() {
-        return 2;
-    }
-}
