+++ app/src/main/res/menu/menu_main.xml
@@ -11,11 +11,6 @@
         android:title="@string/update_quizList_btn"
         app:showAsAction="ifRoom" />
     <item
-        android:icon="@drawable/user_cabinet_icon"
-        android:title="@string/user_cabinet_btn"
-        android:visible="false"
-        app:showAsAction="ifRoom" />
-    <item
         android:id="@+id/menu_item_about"
         android:icon="@android:drawable/ic_menu_info_details"
         android:title="@string/about_menu_item"
