+++ app/src/main/res/values/colors.xml
@@ -5,4 +5,7 @@
     <color name="colorAccent">#63a583</color>
     <color name="color_quiz_item">#FEFFFF</color>
     <color name="main_background">#ADC7CA</color>
+
+    <color name="black">#000000</color>
+    <color name="edit_text_activated">#696666</color>
 </resources>
