+++ app/src/main/res/values/styles.xml
@@ -8,7 +8,11 @@
         <item name="android:textAllCaps">false</item>
     </style>
 
-
+    <style name="MyEditTextTheme" parent="Theme.AppCompat.Light.DarkActionBar">
+        <item name="colorControlNormal">@color/black</item>
+        <item name="colorControlActivated">@color/edit_text_activated</item>
+        <item name="colorControlHighlight">@color/edit_text_activated</item>
+    </style>
 
     <style name="AppTheme.NoActionBar">
         <item name="windowActionBar">false</item>
